requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.25.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0
wandb>=0.15.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
pre-commit>=2.20.0
jupyter>=1.0.0
ipykernel>=6.0.0
